requirements.txt
# Project Genesis-HIV Requirements

# Core Libraries
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Bioinformatics
biopython>=1.79
rdkit>=2022.09.1
openmm>=7.7.0

# AI/Machine Learning
torch>=2.0.0
stable-baselines3>=2.0.0
gymnasium>=0.27.0

# Visualization
streamlit>=1.28.0
plotly>=5.15.0
pyvista>=0.42.0
nglview>=3.0.0

# Data Engine
sqlalchemy>=2.0.0
requests>=2.28.0

# Orchestration
langchain>=0.0.300
crewai>=0.20.0

# Additional utilities
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.1.0